// src/types/formularioDeAva.d.ts
export interface FormularioDeAvaI {
  /** Auto-incrementing primary key */
  id: number;

  createdAt?: string;

  fullName?: string;
  birthDate?: string;
  satisfactionLevel?: string;

  likedMostAndHighlights?: string;
  likedLeastDiscomfort?: string;
  personalChange?: string;
  takeHomeCare?: string;
  tellAFriend?: string;
  praiseOrComplaint?: string;
  improvementSuggestions?: string;
  nextTopics?: string;
  wantsInfo?: string;
  preferredMedia?: string;
  authorizeUse?: string;
}